Add animations to legal pages: Privacy Policy and Terms of Service
- Wrapped PrivacyPolicyContent and TermsOfServiceContent components with LegalPageAnimations for enhanced visual effects.
- Removed unused GSAP animation code from TermsOfServiceContent to streamline the component.
