+++ apps/portal/src/app/(public)/terms-of-service/content.tsx
@@ -1,14 +1,4 @@
-"use client";
-
-import { useRef } from "react";
-import gsap from "gsap";
-import { ScrollTrigger } from "gsap/ScrollTrigger";
-import { useGSAP } from "@gsap/react";
 import Link from "next/link";
-
-if (typeof window !== "undefined") {
-    gsap.registerPlugin(ScrollTrigger);
-}
 
 /* -- Unsplash images -------------------------------------------------------- */
 const img = {
@@ -42,247 +32,15 @@ const tocItems = [
 /* -- Component -------------------------------------------------------------- */
 
 export default function TermsOfServiceContent() {
-    const mainRef = useRef<HTMLElement>(null);
-
-    useGSAP(
-        () => {
-            if (!mainRef.current) return;
-            const prefersReducedMotion = window.matchMedia(
-                "(prefers-reduced-motion: reduce)",
-            ).matches;
-
-            if (prefersReducedMotion) {
-                mainRef.current
-                    .querySelectorAll(".opacity-0")
-                    .forEach((el) => {
-                        (el as HTMLElement).style.opacity = "1";
-                    });
-                return;
-            }
-
-            const $ = (sel: string) =>
-                mainRef.current!.querySelectorAll(sel);
-            const $1 = (sel: string) =>
-                mainRef.current!.querySelector(sel);
-
-            /* -- HERO --------------------------------------------------------- */
-            const heroTl = gsap.timeline({
-                defaults: { ease: "power3.out" },
-            });
-
-            const heroKicker = $1(".hero-kicker");
-            if (heroKicker) {
-                heroTl.fromTo(
-                    heroKicker,
-                    { opacity: 0, y: 20 },
-                    { opacity: 1, y: 0, duration: 0.6 },
-                );
-            }
-
-            const heroWords = $(".hero-headline-word");
-            if (heroWords.length) {
-                heroTl.fromTo(
-                    heroWords,
-                    { opacity: 0, y: 80, rotateX: 40 },
-                    {
-                        opacity: 1,
-                        y: 0,
-                        rotateX: 0,
-                        duration: 1,
-                        stagger: 0.12,
-                    },
-                    "-=0.3",
-                );
-            }
-
-            const heroSubtitle = $1(".hero-subtitle");
-            if (heroSubtitle) {
-                heroTl.fromTo(
-                    heroSubtitle,
-                    { opacity: 0, y: 30 },
-                    { opacity: 1, y: 0, duration: 0.7 },
-                    "-=0.5",
-                );
-            }
-
-            const heroMeta = $(".hero-meta-item");
-            if (heroMeta.length) {
-                heroTl.fromTo(
-                    heroMeta,
-                    { opacity: 0, y: 15 },
-                    { opacity: 1, y: 0, duration: 0.5, stagger: 0.08 },
-                    "-=0.3",
-                );
-            }
-
-            /* Hero image parallax */
-            const heroImgWrap = $1(".hero-img-wrap");
-            if (heroImgWrap) {
-                gsap.fromTo(
-                    heroImgWrap,
-                    { opacity: 0, scale: 1.08 },
-                    {
-                        opacity: 1,
-                        scale: 1,
-                        duration: 1.4,
-                        ease: "power2.out",
-                        delay: 0.2,
-                    },
-                );
-
-                const heroImg = heroImgWrap.querySelector("img");
-                const heroSection = $1(".hero-section");
-                if (heroImg && heroSection) {
-                    gsap.to(heroImg, {
-                        yPercent: 12,
-                        ease: "none",
-                        scrollTrigger: {
-                            trigger: heroSection,
-                            start: "top top",
-                            end: "bottom top",
-                            scrub: true,
-                        },
-                    });
-                }
-            }
-
-            /* -- SECTION REVEALS ---------------------------------------------- */
-            $(".article-block").forEach((block) => {
-                gsap.fromTo(
-                    block,
-                    { opacity: 0, y: 50 },
-                    {
-                        opacity: 1,
-                        y: 0,
-                        duration: 0.8,
-                        ease: "power3.out",
-                        scrollTrigger: {
-                            trigger: block,
-                            start: "top 80%",
-                        },
-                    },
-                );
-            });
-
-            /* -- SPLIT-SCREEN SECTIONS ---------------------------------------- */
-            $(".split-text-left").forEach((el) => {
-                gsap.fromTo(
-                    el,
-                    { opacity: 0, x: -60 },
-                    {
-                        opacity: 1,
-                        x: 0,
-                        duration: 0.8,
-                        ease: "power3.out",
-                        scrollTrigger: {
-                            trigger: el,
-                            start: "top 70%",
-                        },
-                    },
-                );
-            });
-
-            $(".split-img-right").forEach((el) => {
-                gsap.fromTo(
-                    el,
-                    { opacity: 0, x: 60 },
-                    {
-                        opacity: 1,
-                        x: 0,
-                        duration: 0.8,
-                        ease: "power3.out",
-                        scrollTrigger: {
-                            trigger: el,
-                            start: "top 70%",
-                        },
-                    },
-                );
-            });
-
-            $(".split-text-right").forEach((el) => {
-                gsap.fromTo(
-                    el,
-                    { opacity: 0, x: 60 },
-                    {
-                        opacity: 1,
-                        x: 0,
-                        duration: 0.8,
-                        ease: "power3.out",
-                        scrollTrigger: {
-                            trigger: el,
-                            start: "top 70%",
-                        },
-                    },
-                );
-            });
-
-            $(".split-img-left").forEach((el) => {
-                gsap.fromTo(
-                    el,
-                    { opacity: 0, x: -60 },
-                    {
-                        opacity: 1,
-                        x: 0,
-                        duration: 0.8,
-                        ease: "power3.out",
-                        scrollTrigger: {
-                            trigger: el,
-                            start: "top 70%",
-                        },
-                    },
-                );
-            });
-
-            /* -- PULL QUOTES -------------------------------------------------- */
-            $(".pull-quote-block").forEach((quote) => {
-                gsap.fromTo(
-                    quote,
-                    { opacity: 0, scale: 0.96 },
-                    {
-                        opacity: 1,
-                        scale: 1,
-                        duration: 0.9,
-                        ease: "power3.out",
-                        scrollTrigger: {
-                            trigger: quote,
-                            start: "top 80%",
-                        },
-                    },
-                );
-            });
-
-            /* -- CTA ---------------------------------------------------------- */
-            const ctaContent = $1(".final-cta-content");
-            const ctaSection = $1(".final-cta");
-            if (ctaContent && ctaSection) {
-                gsap.fromTo(
-                    ctaContent,
-                    { opacity: 0, y: 50 },
-                    {
-                        opacity: 1,
-                        y: 0,
-                        duration: 1,
-                        ease: "power3.out",
-                        scrollTrigger: {
-                            trigger: ctaSection,
-                            start: "top 80%",
-                        },
-                    },
-                );
-            }
-        },
-        { scope: mainRef },
-    );
-
     return (
-        <main ref={mainRef} className="overflow-hidden">
+        <main className="overflow-hidden">
             {/* =================================================================
                 HERO -- Split-screen 60/40 with diagonal clip
                ================================================================= */}
             <section className="hero-section relative min-h-[92vh] flex items-center bg-base-100">
                 {/* Right image panel */}
                 <div
-                    className="hero-img-wrap absolute inset-0 lg:left-[58%] opacity-0"
+                    className="hero-img-wrap absolute inset-0 lg:left-[58%]"
                     style={{
                         clipPath:
                             "polygon(8% 0, 100% 0, 100% 100%, 0% 100%)",
@@ -299,36 +57,36 @@ export default function TermsOfServiceContent() {
                 {/* Content panel -- 60% on desktop */}
                 <div className="relative z-10 container mx-auto px-6 lg:px-12 py-28">
                     <div className="max-w-2xl">
-                        <p className="hero-kicker text-sm font-semibold uppercase tracking-[0.2em] text-secondary mb-6 opacity-0">
+                        <p className="hero-kicker text-sm font-semibold uppercase tracking-[0.2em] text-secondary mb-6">
                             <i className="fa-duotone fa-regular fa-file-contract mr-2"></i>
                             Legal Agreement
                         </p>
 
                         <h1 className="text-4xl md:text-5xl lg:text-6xl xl:text-7xl font-black leading-[0.92] tracking-tight mb-8">
-                            <span className="hero-headline-word inline-block opacity-0 text-base-content lg:text-base-content">
+                            <span className="hero-headline-word inline-block text-base-content lg:text-base-content">
                                 Terms of
                             </span>{" "}
-                            <span className="hero-headline-word inline-block opacity-0 text-secondary">
+                            <span className="hero-headline-word inline-block text-secondary">
                                 Service.
                             </span>
                         </h1>
 
-                        <p className="hero-subtitle text-lg md:text-xl text-base-content/70 leading-relaxed max-w-xl mb-8 opacity-0">
+                        <p className="hero-subtitle text-lg md:text-xl text-base-content/70 leading-relaxed max-w-xl mb-8">
                             The rules, guidelines, and mutual commitments that
                             govern your use of the Splits Network platform.
                             Please read carefully before using our services.
                         </p>
 
                         <div className="flex flex-wrap items-center gap-6">
-                            <span className="hero-meta-item text-xs uppercase tracking-[0.15em] text-base-content/50 opacity-0">
+                            <span className="hero-meta-item text-xs uppercase tracking-[0.15em] text-base-content/50">
                                 <i className="fa-duotone fa-regular fa-building mr-1"></i>
                                 Employment Networks, Inc.
                             </span>
-                            <span className="hero-meta-item text-xs uppercase tracking-[0.15em] text-base-content/50 opacity-0">
+                            <span className="hero-meta-item text-xs uppercase tracking-[0.15em] text-base-content/50">
                                 <i className="fa-duotone fa-regular fa-calendar mr-1"></i>
                                 Last Updated: February 20, 2026
                             </span>
-                            <span className="hero-meta-item text-xs uppercase tracking-[0.15em] text-base-content/50 opacity-0">
+                            <span className="hero-meta-item text-xs uppercase tracking-[0.15em] text-base-content/50">
                                 <i className="fa-duotone fa-regular fa-clock mr-1"></i>
                                 16 Sections
                             </span>
@@ -342,7 +100,7 @@ export default function TermsOfServiceContent() {
                ================================================================= */}
             <section className="py-28 bg-base-100">
                 <div className="container mx-auto px-6 lg:px-12">
-                    <div className="article-block max-w-3xl mx-auto opacity-0">
+                    <div className="article-block max-w-3xl mx-auto">
                         <p className="text-sm font-semibold uppercase tracking-[0.2em] text-secondary mb-4">
                             <i className="fa-duotone fa-regular fa-list mr-2"></i>
                             Quick Navigation
@@ -378,7 +136,7 @@ export default function TermsOfServiceContent() {
                ================================================================= */}
             <section id="acceptance" className="py-28 bg-base-200">
                 <div className="container mx-auto px-6 lg:px-12">
-                    <div className="article-block max-w-3xl mx-auto opacity-0">
+                    <div className="article-block max-w-3xl mx-auto">
                         <p className="text-sm font-semibold uppercase tracking-[0.2em] text-secondary mb-4">
                             01 -- Agreement
                         </p>
@@ -425,7 +183,7 @@ export default function TermsOfServiceContent() {
                 <div className="container mx-auto px-6 lg:px-12">
                     <div className="grid lg:grid-cols-5 gap-12 lg:gap-16 items-center">
                         {/* Text -- 3 of 5 columns (60%) */}
-                        <div className="split-text-left lg:col-span-3 opacity-0">
+                        <div className="split-text-left lg:col-span-3">
                             <p className="text-sm font-semibold uppercase tracking-[0.2em] text-secondary mb-4">
                                 02 -- The Platform
                             </p>
@@ -468,7 +226,7 @@ export default function TermsOfServiceContent() {
                         </div>
 
                         {/* Image -- 2 of 5 columns (40%) */}
-                        <div className="split-img-right lg:col-span-2 opacity-0">
+                        <div className="split-img-right lg:col-span-2">
                             <div
                                 className="relative overflow-hidden"
                                 style={{
@@ -494,7 +252,7 @@ export default function TermsOfServiceContent() {
                ================================================================= */}
             <section id="eligibility" className="py-28 bg-base-200">
                 <div className="container mx-auto px-6 lg:px-12">
-                    <div className="article-block max-w-3xl mx-auto opacity-0">
+                    <div className="article-block max-w-3xl mx-auto">
                         <p className="text-sm font-semibold uppercase tracking-[0.2em] text-secondary mb-4">
                             03 -- Requirements
                         </p>
@@ -553,7 +311,7 @@ export default function TermsOfServiceContent() {
                ================================================================= */}
             <section id="accounts" className="py-28 bg-base-100">
                 <div className="container mx-auto px-6 lg:px-12">
-                    <div className="article-block max-w-3xl mx-auto opacity-0">
+                    <div className="article-block max-w-3xl mx-auto">
                         <p className="text-sm font-semibold uppercase tracking-[0.2em] text-secondary mb-4">
                             04 -- Your Account
                         </p>
@@ -627,7 +385,7 @@ export default function TermsOfServiceContent() {
             {/* =================================================================
                 PULL QUOTE 1
                ================================================================= */}
-            <section className="pull-quote-block py-20 bg-neutral text-neutral-content opacity-0">
+            <section className="pull-quote-block py-20 bg-neutral text-neutral-content">
                 <div className="container mx-auto px-6 lg:px-12">
                     <div className="max-w-4xl mx-auto">
                         <div className="border-l-4 border-secondary pl-8 lg:pl-12">
@@ -651,7 +409,7 @@ export default function TermsOfServiceContent() {
                ================================================================= */}
             <section id="conduct" className="py-28 bg-base-200">
                 <div className="container mx-auto px-6 lg:px-12">
-                    <div className="article-block max-w-3xl mx-auto opacity-0">
+                    <div className="article-block max-w-3xl mx-auto">
                         <p className="text-sm font-semibold uppercase tracking-[0.2em] text-secondary mb-4">
                             05 -- Expectations
                         </p>
@@ -712,7 +470,7 @@ export default function TermsOfServiceContent() {
                 <div className="container mx-auto px-6 lg:px-12">
                     <div className="grid lg:grid-cols-5 gap-12 lg:gap-16 items-center">
                         {/* Image -- 2 of 5 columns (40%) */}
-                        <div className="split-img-left lg:col-span-2 opacity-0">
+                        <div className="split-img-left lg:col-span-2">
                             <div
                                 className="relative overflow-hidden"
                                 style={{
@@ -729,7 +487,7 @@ export default function TermsOfServiceContent() {
                         </div>
 
                         {/* Text -- 3 of 5 columns (60%) */}
-                        <div className="split-text-right lg:col-span-3 opacity-0">
+                        <div className="split-text-right lg:col-span-3">
                             <p className="text-sm font-semibold uppercase tracking-[0.2em] text-secondary mb-4">
                                 06 -- Rules of Engagement
                             </p>
@@ -809,7 +567,7 @@ export default function TermsOfServiceContent() {
                ================================================================= */}
             <section id="fees" className="py-28 bg-base-200">
                 <div className="container mx-auto px-6 lg:px-12">
-                    <div className="article-block max-w-3xl mx-auto opacity-0">
+                    <div className="article-block max-w-3xl mx-auto">
                         <p className="text-sm font-semibold uppercase tracking-[0.2em] text-secondary mb-4">
                             07 -- Financial Terms
                         </p>
@@ -883,7 +641,7 @@ export default function TermsOfServiceContent() {
                ================================================================= */}
             <section id="ip" className="py-28 bg-base-100">
                 <div className="container mx-auto px-6 lg:px-12">
-                    <div className="article-block max-w-3xl mx-auto opacity-0">
+                    <div className="article-block max-w-3xl mx-auto">
                         <p className="text-sm font-semibold uppercase tracking-[0.2em] text-secondary mb-4">
                             08 -- Ownership
                         </p>
@@ -945,7 +703,7 @@ export default function TermsOfServiceContent() {
             {/* =================================================================
                 PULL QUOTE 2
                ================================================================= */}
-            <section className="pull-quote-block py-20 bg-secondary text-secondary-content opacity-0">
+            <section className="pull-quote-block py-20 bg-secondary text-secondary-content">
                 <div className="container mx-auto px-6 lg:px-12">
                     <div className="max-w-4xl mx-auto text-center">
                         <i className="fa-duotone fa-regular fa-quote-left text-4xl text-secondary-content/20 mb-6 block"></i>
@@ -966,7 +724,7 @@ export default function TermsOfServiceContent() {
                ================================================================= */}
             <section id="ai-features" className="py-28 bg-base-200">
                 <div className="container mx-auto px-6 lg:px-12">
-                    <div className="article-block max-w-3xl mx-auto opacity-0">
+                    <div className="article-block max-w-3xl mx-auto">
                         <p className="text-sm font-semibold uppercase tracking-[0.2em] text-secondary mb-4">
                             09 -- Artificial Intelligence
                         </p>
@@ -1058,7 +816,7 @@ export default function TermsOfServiceContent() {
                ================================================================= */}
             <section id="privacy" className="py-28 bg-base-100">
                 <div className="container mx-auto px-6 lg:px-12">
-                    <div className="article-block max-w-3xl mx-auto opacity-0">
+                    <div className="article-block max-w-3xl mx-auto">
                         <p className="text-sm font-semibold uppercase tracking-[0.2em] text-secondary mb-4">
                             10 -- Data Protection
                         </p>
@@ -1129,7 +887,7 @@ export default function TermsOfServiceContent() {
             {/* =================================================================
                 PULL QUOTE 3
                ================================================================= */}
-            <section className="pull-quote-block py-20 bg-neutral text-neutral-content opacity-0">
+            <section className="pull-quote-block py-20 bg-neutral text-neutral-content">
                 <div className="container mx-auto px-6 lg:px-12">
                     <div className="max-w-4xl mx-auto">
                         <div className="border-l-4 border-secondary pl-8 lg:pl-12">
@@ -1152,7 +910,7 @@ export default function TermsOfServiceContent() {
                ================================================================= */}
             <section id="disclaimers" className="py-28 bg-base-200">
                 <div className="container mx-auto px-6 lg:px-12">
-                    <div className="article-block max-w-3xl mx-auto opacity-0">
+                    <div className="article-block max-w-3xl mx-auto">
                         <p className="text-sm font-semibold uppercase tracking-[0.2em] text-secondary mb-4">
                             11 -- Legal Protections
                         </p>
@@ -1231,7 +989,7 @@ export default function TermsOfServiceContent() {
                ================================================================= */}
             <section id="indemnification" className="py-28 bg-base-100">
                 <div className="container mx-auto px-6 lg:px-12">
-                    <div className="article-block max-w-3xl mx-auto opacity-0">
+                    <div className="article-block max-w-3xl mx-auto">
                         <p className="text-sm font-semibold uppercase tracking-[0.2em] text-secondary mb-4">
                             12 -- Your Responsibility
                         </p>
@@ -1284,7 +1042,7 @@ export default function TermsOfServiceContent() {
                ================================================================= */}
             <section id="termination" className="py-28 bg-base-200">
                 <div className="container mx-auto px-6 lg:px-12">
-                    <div className="article-block max-w-3xl mx-auto opacity-0">
+                    <div className="article-block max-w-3xl mx-auto">
                         <p className="text-sm font-semibold uppercase tracking-[0.2em] text-secondary mb-4">
                             13 -- Account Closure
                         </p>
@@ -1361,7 +1119,7 @@ export default function TermsOfServiceContent() {
                ================================================================= */}
             <section id="disputes" className="py-28 bg-base-100">
                 <div className="container mx-auto px-6 lg:px-12">
-                    <div className="article-block max-w-3xl mx-auto opacity-0">
+                    <div className="article-block max-w-3xl mx-auto">
                         <p className="text-sm font-semibold uppercase tracking-[0.2em] text-secondary mb-4">
                             14 -- Resolving Disagreements
                         </p>
@@ -1439,7 +1197,7 @@ export default function TermsOfServiceContent() {
                ================================================================= */}
             <section id="governing-law" className="py-28 bg-base-200">
                 <div className="container mx-auto px-6 lg:px-12">
-                    <div className="article-block max-w-3xl mx-auto opacity-0">
+                    <div className="article-block max-w-3xl mx-auto">
                         <p className="text-sm font-semibold uppercase tracking-[0.2em] text-secondary mb-4">
                             15 -- Jurisdiction
                         </p>
@@ -1478,7 +1236,7 @@ export default function TermsOfServiceContent() {
                ================================================================= */}
             <section id="changes" className="py-28 bg-base-100">
                 <div className="container mx-auto px-6 lg:px-12">
-                    <div className="article-block max-w-3xl mx-auto opacity-0">
+                    <div className="article-block max-w-3xl mx-auto">
                         <p className="text-sm font-semibold uppercase tracking-[0.2em] text-secondary mb-4">
                             16 -- Updates
                         </p>
@@ -1536,7 +1294,7 @@ export default function TermsOfServiceContent() {
                ================================================================= */}
             <section className="py-28 bg-base-200">
                 <div className="container mx-auto px-6 lg:px-12">
-                    <div className="article-block max-w-3xl mx-auto opacity-0">
+                    <div className="article-block max-w-3xl mx-auto">
                         <p className="text-sm font-semibold uppercase tracking-[0.2em] text-secondary mb-4">
                             Related Policies
                         </p>
@@ -1590,7 +1348,7 @@ export default function TermsOfServiceContent() {
                ================================================================= */}
             <section className="final-cta py-28 bg-secondary text-secondary-content">
                 <div className="container mx-auto px-6 lg:px-12">
-                    <div className="final-cta-content max-w-4xl mx-auto text-center opacity-0">
+                    <div className="final-cta-content max-w-4xl mx-auto text-center">
                         <p className="text-sm font-semibold uppercase tracking-[0.2em] text-secondary-content/50 mb-6">
                             <i className="fa-duotone fa-regular fa-envelope mr-2"></i>
                             Contact Us
