+++ apps/portal/src/app/(public)/terms-of-service/page.tsx
@@ -1,5 +1,6 @@
 import { Metadata } from "next";
 import { buildCanonical } from "@/lib/seo";
+import LegalPageAnimations from "../legal-page-animations";
 import TermsOfServiceContent from "./content";
 
 export const metadata: Metadata = {
@@ -16,5 +17,9 @@ export const metadata: Metadata = {
 };
 
 export default function TermsOfServicePage() {
-    return <TermsOfServiceContent />;
+    return (
+        <LegalPageAnimations>
+            <TermsOfServiceContent />
+        </LegalPageAnimations>
+    );
 }
